Fix testhost depproj chmod when win host restored
When building on Linux, this Exec assumes a Linux configuration is
restored so the testhost package it picks up will have dotnet. This
fails with '-allconfigurations' when a Windows configuration is used.
This fix might not be the best, it only fixes the assumption.

Tracked by https://github.com/dotnet/corefx/issues/34183