Add BuildsOnUnsupportedRIDs to prevent UAP runtime
Setting this on the NETCoreApp package continues the behavior where
a runtime package is always built, even for an unrecognized
PackageRID. However, the UAP project doesn't build natives on
arbitrary Linuxes, so disable this behavior for that package.

The property name probably needs to be made more specific to this
scenario, or perhaps the UAP package build could be fixed.